HoldFps int to double
15.07.26	- HoldFps - change from int to double

+++ SPOUTSDK/SpoutGL/SpoutReceiver.cpp
@@ -91,6 +91,7 @@
 //		13.10.25	- Add GetSenderTexture, GetSenderTexture
 //					  replace GetSpoutVersion with GetSDKversion
 //		22.01.26	- Review - update copyright year
+//		15.07.26	- HoldFps - change from int to double
 //
 // ====================================================================================
 //
@@ -306,7 +307,7 @@ bool SpoutReceiver::IsFrameCountEnabled()
 }
 
 //---------------------------------------------------------
-void SpoutReceiver::HoldFps(int fps)
+void SpoutReceiver::HoldFps(double fps)
 {
 	spout.HoldFps(fps);
 }
